separated dj and recommendation, both working
